phase4/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.0.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
reportlab>=4.0.0
google-genai>=1.0.0
easyocr
numpy
